backend/requirements.txt
asgiref==3.8.1
boto3==1.34.74
botocore==1.34.74
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
coverage==7.4.4
cryptography==42.0.5
decorator==4.4.2
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.7
django-anymail==10.3
django-ckeditor-5==0.2.12
django-cors-headers==4.3.1
django-jazzmin==2.6.1
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environs==11.0.0
filelock==3.13.3
gunicorn==21.2.0
identify==2.5.35
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
inflection==0.5.1
jmespath==1.0.1
marshmallow==3.21.1
mccabe==0.7.0
moviepy==1.0.3
nodeenv==1.8.0
numpy==1.26.4
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.7.0
proglog==0.1.10
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
stripe==8.9.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
